jokaisessa DAO-luokassa on nyt myos metodi jolla tietokantaan voi lisata riveja, id:t ja timestampit tulevat niissa automaattisesti.tein myos bugikorjauksia.

+++ src/main/java/tikape/runko/database/AlueDao.java
@@ -32,12 +32,12 @@ public class AlueDao implements Dao<Alue, Integer> {
         if (!hasOne) {
             return null;
         }
-        
+
         int alueenid = rs.getInt("id");
         String alueennimi = rs.getString("nimi");
         String kuvaus = rs.getString("kuvaus");
 
-        Alue a = new Alue(alueenid, alueennimi,kuvaus);
+        Alue a = new Alue(alueenid, alueennimi, kuvaus);
 
         rs.close();
         stmt.close();
@@ -46,11 +46,11 @@ public class AlueDao implements Dao<Alue, Integer> {
         return a;
     }
 
-    
     /**
      * tämä metodi palauttaa listan kaikista tietokannan Alueista
+     *
      * @return lista kaikista alueista
-     * @throws SQLException 
+     * @throws SQLException
      */
     @Override
     public List<Alue> findAll() throws SQLException {
@@ -77,14 +77,42 @@ public class AlueDao implements Dao<Alue, Integer> {
 
     @Override
     public void delete(Integer id) throws SQLException {
-        // ei toteutettu alkuperäisessä opiskelijaDaossa
-        Connection connection = database.getConnection();
-        PreparedStatement stmt = connection.prepareStatement("DELETE FROM Alue WHERE id = ?");
-        //vaihdetaan kysymysmerkki edellisellä rivillä parametrissa saatuun kokonaislukuun
-        stmt.setObject(1, id);
-        //toteutetaan SQL query
-        stmt.executeQuery();
+        //ei tee mitään jos paramissa annettua id:tä ei löydy
+        if (this.findOne(id) != null) {
+            Connection connection = database.getConnection();
+            PreparedStatement stmt = connection.prepareStatement("DELETE FROM Alue WHERE id = ?");
+            //vaihdetaan kysymysmerkki edellisellä rivillä parametrissa saatuun kokonaislukuun
+            stmt.setObject(1, id);
+            //toteutetaan SQL query
+            stmt.executeQuery();
+            stmt.close();
+            connection.close();
+
+        }
     }
 
+    public Alue lisaa(String nimi, String kuvaus) throws SQLException {
+        Connection connection = database.getConnection();
+        //valitaan suurin numero id sarakkeessa
+        PreparedStatement stmt = connection.prepareStatement("SELECT max(id) FROM Alue");
+        ResultSet rs = stmt.executeQuery();
+        //jos taulu on tyhjä, ensimmäisen rivin id on 0
+        int uudenAlueenId = 0;
+        //muuten inkrementoidaan yhdellä seuraavaan id lukuun
+        if (rs.next()) {
+            uudenAlueenId = rs.getInt(1) + 1;
+        }
+        //Lopuksi tehdään SQL update
+        PreparedStatement stmt2 = connection.prepareStatement("INSERT INTO Alue VALUES(?,?,?)");
+        stmt2.setInt(1, uudenAlueenId);
+        stmt2.setString(2, nimi);
+        stmt2.setString(3, kuvaus);
+        stmt2.executeUpdate();
+        rs.close();
+        stmt.close();
+        stmt2.close();
+        connection.close();
+        return this.findOne(uudenAlueenId);
+    }
 
 }
